+++ BE/.prettierrc
@@ -8,5 +8,6 @@
   "bracketSpacing": true,
   "jsxSingleQuote": true,
   "arrowParens": "always",
-  "endOfLine": "auto"
+  "endOfLine": "auto",
+  "bracketSameLine": true
 }
