+++ BE/src/game-user/game-user.service.ts
@@ -1,10 +1,10 @@
 import { FindGameUserUsecase } from './usecase/find.game-user.usecase';
 import { RegisterGameUserUsecase } from './usecase/register.game-user.usecase';
 import { Inject, Injectable } from '@nestjs/common';
-import { GameHistoryEntity } from '../game/entity/game.history.entity';
+import { GameHistoryEntity } from '../game/entity/game-history.entity';
 import { FIND_USER_USECASE, FindUserUsecase } from '../user/usecase/find.user.usecase';
 import { GAME_USER_REPOSITORY, GameUserRepository } from './repository/game-user.repository';
-import { GameUserEntity } from './enitity/game.user.entity';
+import { GameUserEntity } from './enitity/game-user.entity';
 import { FindUserRequest } from '../user/dto/find-user.request';
 import { Transactional } from 'typeorm-transactional';
