+++ BE/src/game-user/usecase/find.game-user.usecase.ts
@@ -1,4 +1,4 @@
-import { GameUserEntity } from '../enitity/game.user.entity';
+import { GameUserEntity } from '../enitity/game-user.entity';
 
 export const FIND_GAME_USER_USECASE = Symbol('FIND_GAME_USER_USECASE');
