+++ BE/src/game/entity/game-history.result.ts
