+++ BE/src/game/mafia-role.ts
