+++ BE/src/game/usecase/allocate.user-role.usecase.ts
